Got motor control funcitons working

+++ include/micromouse/core/pid.h
@@ -22,29 +22,29 @@ extern "C" {
 /**
  * The P term in a PID controller.
  */
-double p_term(
-    const signed int err, 
-    const double kp
+long double p_term(
+    const long double err, 
+    const long double kp
 );
 
 /**
  * The I term in a PID controller.
  */
-double i_term(
-    const signed int err, 
-    const double integral, 
-    const unsigned int dt, 
-    const double ki
+long double i_term(
+    const long double err, 
+    const long double integral, 
+    const long double dt, 
+    const long double ki
 );
 
 /**
  * The D term in a PID controller.
  */
-double d_term(
-    const signed int err, 
-    const signed int prev_err, 
-    const unsigned int dt, 
-    const double kd
+long double d_term(
+    const long double err, 
+    const long double prev_err, 
+    const long double dt, 
+    const long double kd
 );
